chore: updated GitHub Action

+++ apps/web/global.teardown.ts
@@ -1,6 +1,10 @@
 import { deleteUser } from '@documenso/lib/server-only/user/delete-user';
 
 async function teardown() {
+  if (!process.env.E2E_TEST_USERNAME) {
+    return;
+  }
+
   try {
     await deleteUser(process.env.E2E_TEST_USERNAME);
   } catch (e) {
